requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
pyaml==21.10.1
PyYAML==6.0
requests==2.28.0
urllib3==1.26.9
